% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoloc.R
\name{query_geoloc}
\alias{query_geoloc}
\title{Query the Geolocation Service}
\usage{
query_geoloc(query = "Baleine Rd, Baleine, NS", expand = NULL,
  callback = NULL, verb = c("locate", "suggest", "autocomplete")[1],
  baseuri = "http://geogratis.gc.ca/services/geolocation/en", parse = TRUE,
  ...)
}
\arguments{
\item{query}{character vector of one or more locations to query}

\item{expand}{character vector or NULL to retrieve additional bits of info per object.
By default NULL, but allowed is \code{'score,component'} or one or the other}

\item{callback}{NULL see \url{http://geogratis.gc.ca/site/eng/geoloc}}

\item{verb}{character the resource to query, by default 'locate'}

\item{baseuri}{character the resource uri without the verb}

\item{parse}{logical if TRUE then try to parse to sf dataframe}

\item{...}{further arguments for \code{parse_geoloc()}}
}
\description{
Query the Geolocation Service
}
\examples{
\dontrun{
   # query one of the prettiest places on the planet
   x <- query_geoloc("Baleine Rd, Baleine, NS")
}
}
\seealso{
\url{http://geogratis.gc.ca/site/eng/geoloc}
}
